mcp-gemma/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-gemma"
version = "1.0.0"
description = "High-performance Model Context Protocol server for gemma.cpp integration"
authors = [
    {name = "LLM Development Team", email = "dev@example.com"}
]
readme = "README.md"
license = {file = "LICENSE"}
keywords = ["mcp", "gemma", "llm", "ai", "server", "high-performance"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.11"
dependencies = [
    # Web frameworks and networking - optimized versions
    "aiohttp[speedups]>=3.9.0",
    "aiohttp-cors>=0.7.0",
    "websockets>=12.0",

    # Data processing and validation - Pydantic v2
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",

    # High-performance JSON and serialization
    "orjson>=3.9.0",
    "msgpack>=1.0.5",

    # Redis with async support
    "redis[hiredis]>=5.0.0",

    # Configuration management
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",

    # Performance optimizations
    "uvloop>=0.19.0;sys_platform!='win32'",

    # Async utilities
    "asyncio-throttle>=1.0.2",
    "aiofiles>=23.2.1",

    # Machine learning and text processing
    "sentence-transformers>=2.5.0",
    "numpy>=1.24.0",

    # Caching and memory optimization
    "cachetools>=5.3.0",
    "psutil>=5.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-benchmark>=4.0.0",
    "black>=23.11.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "pre-commit>=3.5.0",
    "types-redis>=4.6.0",
]
enhanced = [
    "faiss-cpu>=1.7.4",
    "transformers>=4.35.0",
    "torch>=2.1.0",
    "spacy>=3.7.0",
    "scikit-learn>=1.3.0",
]
monitoring = [
    "opentelemetry-api>=1.21.0",
    "opentelemetry-sdk>=1.21.0",
    "opentelemetry-instrumentation>=0.42b0",
    "prometheus-client>=0.19.0",
    "structlog>=23.2.0",
]
performance = [
    "orjson>=3.9.0",
    "msgpack>=1.0.5",
    "lz4>=4.3.0",
    "brotli>=1.1.0",
    "uvloop>=0.19.0;sys_platform!='win32'",
    "asyncpg>=0.29.0",
    "httptools>=0.6.0",
    "ujson>=5.8.0",
]

[project.scripts]
mcp-gemma = "server.main:main"
mcp-gemma-client = "client.cli:main"

[project.urls]
Homepage = "https://github.com/llm-dev/mcp-gemma"
Documentation = "https://mcp-gemma.readthedocs.io"
Repository = "https://github.com/llm-dev/mcp-gemma.git"
Issues = "https://github.com/llm-dev/mcp-gemma/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["server*", "client*", "integration*"]
exclude = ["tests*", "docs*", "scripts*"]

[tool.black]
line-length = 100
target-version = ['py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
target-version = "py311"
line-length = 100
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "C",    # flake8-comprehensions
    "B",    # flake8-bugbear
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "RUF",  # Ruff-specific rules
    "ASYNC", # async-related rules
    "PERF", # performance-related rules
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
    "RUF012", # Mutable class attributes should be annotated with `typing.ClassVar`
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.11"
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
strict_equality = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "redis.*",
    "sentence_transformers.*",
    "websockets.*",
    "aiohttp.*",
    "orjson.*",
    "msgpack.*",
    "uvloop.*",
]
ignore_missing_imports = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

[tool.pytest.ini_options]
minversion = "7.4"
addopts = "-ra -q --strict-markers --strict-config --benchmark-disable"
testpaths = ["tests"]
filterwarnings = [
    "error",
    "ignore::UserWarning",
    "ignore::DeprecationWarning",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "benchmark: marks tests as benchmarks",
]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["server", "client", "integration"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
